Refactor efficientPower function to use long long instead of double for temp variable

+++ GFG/power.cpp
@@ -55,7 +55,7 @@ Example how this works:
 */
 long long efficientPower(long long x, long long n) {
     if(n==0) { return 1;}
-    double temp=pow(x,n/2);
+    long long temp=pow(x,n/2);
     if (n%2==0){return temp*temp;}
     else {return x*temp*temp;}
 }
